feat(laminar-agent): edits round 3 — header actions, file splits, zustand fix
- Move New Chat button to header (reload, change view, close icons)
- Split tool-call-cards into one component per file
- Fix deprecated Zustand useStore signature with useShallow
- Remove resolved TODO/QUESTION comments
