+++ frontend/components/laminar-agent/tool-call-cards/sql-tool-card.tsx
@@ -1,28 +1,10 @@
 "use client";
 
-import { ChevronDown, ChevronRight, Database, ListTree, Loader2 } from "lucide-react";
+import { ChevronDown, ChevronRight, Database, Loader2 } from "lucide-react";
 import { useState } from "react";
 
 import CodeHighlighter from "@/components/ui/code-highlighter";
 import { CopyButton } from "@/components/ui/copy-button";
-
-interface CompactTraceCardProps {
-  isLoading?: boolean;
-}
-
-export function CompactTraceCard({ isLoading }: CompactTraceCardProps) {
-  return (
-    <div className="bg-muted/50 rounded-lg p-3 border">
-      <div className="flex items-center gap-2">
-        <ListTree className="size-3.5 text-muted-foreground shrink-0" />
-        <span className="text-xs font-medium text-muted-foreground">
-          {isLoading ? "Fetching trace context..." : "Fetched trace context"}
-        </span>
-        {isLoading && <Loader2 className="size-3 animate-spin text-muted-foreground" />}
-      </div>
-    </div>
-  );
-}
 
 interface SqlToolCardProps {
   query: string;
@@ -36,7 +18,7 @@ export function SqlToolCard({ query, isLoading }: SqlToolCardProps) {
     <div className="bg-muted/50 rounded-lg border">
       <button
         onClick={() => setExpanded(!expanded)}
-        className="flex items-center gap-2 w-full p-3 text-left hover:bg-muted/80 transition-colors rounded-lg"
+        className="flex items-center gap-2 w-full px-3 py-2 text-left hover:bg-muted/80 transition-colors rounded-lg"
       >
         <Database className="size-3.5 text-muted-foreground shrink-0" />
         <span className="text-xs font-medium text-muted-foreground flex-1">
