+++ frontend/components/traces/trace-view/header/signals-pill.tsx
@@ -1,9 +1,9 @@
 "use client";
 
-import { ExternalLink, Loader, Sparkles, Zap } from "lucide-react";
+import { ExternalLink, Loader, Radio, Sparkles } from "lucide-react";
 import { useParams } from "next/navigation";
 import { useCallback, useEffect, useState } from "react";
-import { shallow } from "zustand/shallow";
+import { useShallow } from "zustand/react/shallow";
 
 import { useLaminarAgentStore } from "@/components/laminar-agent/store";
 import { Button } from "@/components/ui/button";
@@ -37,13 +37,12 @@ export default function SignalsPill({ traceId }: SignalsPillProps) {
   const [isLoading, setIsLoading] = useState(true);
 
   const { setViewMode, setPrefillInput, viewMode, aiEnabled } = useLaminarAgentStore(
-    (s) => ({
+    useShallow((s) => ({
       setViewMode: s.setViewMode,
       setPrefillInput: s.setPrefillInput,
       viewMode: s.viewMode,
       aiEnabled: s.aiEnabled,
-    }),
-    shallow
+    }))
   );
 
   useEffect(() => {
@@ -93,7 +92,7 @@ export default function SignalsPill({ traceId }: SignalsPillProps) {
 
   const handleExplainSignal = useCallback(
     (event: SignalEvent) => {
-      const prompt = `Show me the payload of this signal event ${event.id}, explain why it was detected on this trace ${traceId}, and detail which spans are relevant and why`;
+      const prompt = `Show me the payload of this signal event ${event.id} formatted in a table, explain why it was detected on this trace ${traceId}, and detail which spans are relevant and why`;
       setPrefillInput(prompt);
       if (viewMode === "collapsed") {
         setViewMode("floating");
@@ -126,8 +125,8 @@ export default function SignalsPill({ traceId }: SignalsPillProps) {
   return (
     <DropdownMenu>
       <DropdownMenuTrigger asChild>
-        <Button variant="outline" className="h-6 text-xs px-1.5 border-primary text-primary hover:bg-primary/10">
-          <Zap size={14} />
+        <Button variant="outline" className="h-6 text-xs px-1.5 border-primary text-primary hover:bg-primary/10 gap-1">
+          <Radio size={14} />
           {events.length} {events.length === 1 ? "signal" : "signals"}
         </Button>
       </DropdownMenuTrigger>
